requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-dropbox==0.0.2
django-rest-auth==0.9.5
djangorestframework==3.12.2
dropbox==11.5.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
ply==3.11
psycopg2==2.7.7
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stone==3.2.1
urllib3==1.26.3
whitenoise==5.2.0
tika==1.24
epub2txt==0.1.1